Add mapping in org noter
Maps `org-noter-insert-note` to be `i` and
`org-noter-insert-note-toggle-no-questions` to be `I` when
`org-noter-doc-mode` is enabled. This is because the mapping is
curiously missing without it and only the
`org-noter-insert-precise-note` is mapped to `M-i`.

This commit also includes a fix for `mu4e` which made many emails occur
twice or more and also did not allow them to be handled correctly. It's
my suspicion that this is because of the `![Gmail]*` line in the
`mbsyncrc`.